SAROJ-1 Split in multiple files based on function
