+++ WebServiceModules/TextExtractor/textExtractor_XMLParser.py
